refactoring: fix project structure; add gradle; add missing classes
