--- META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Rsrc-Class-Path: ./ commons-digester-2.1.jar bcprov-jdk14-138.jar cast
- or-1.2.jar jfreechart-1.0.12.jar jcommon-1.0.15.jar poi-ooxml-3.7.jar
-  xmlbeans-2.3.0.jar commons-beanutils-1.8.0.jar bcmail-jdk14-138.jar 
- serializer-2.7.1.jar xalan-2.7.1.jar stax-api-1.0.1.jar commons-loggi
- ng-1.1.1.jar commons-collections-2.1.jar geronimo-stax-api_1.0_spec-1
- .0.jar bctsp-jdk14-1.38.jar xml-apis-1.3.02.jar jackson-core-lgpl-1.7
- .4.jar commons-io-2.1.jar itext-2.1.7.jar jackson-mapper-lgpl-1.7.4.j
- ar bcmail-jdk14-1.38.jar jasperreports-4.5.1.jar bcprov-jdk14-1.38.ja
- r jdtcore-3.1.0.jar poi-3.7.jar dom4j-1.6.1.jar poi-ooxml-schemas-3.7
- .jar
-Class-Path: .
-Rsrc-Main-Class: com.microting.report.jasper.JasperExporter
-Main-Class: org.eclipse.jdt.internal.jarinjarloader.JarRsrcLoader
-
