+++ build.gradle
@@ -1,0 +1,22 @@
+plugins {
+    id 'java'
+    id 'application'
+}
+
+mainClassName = 'com.microting.report.jasper.JasperExporter'
+
+
+repositories {
+    jcenter()
+
+    maven { url "http://jasperreports.sourceforge.net/maven2" }
+    maven { url "http://jaspersoft.artifactoryonline.com/jaspersoft/third-party-ce-artifacts" }
+}
+
+dependencies {
+    compile 'com.google.guava:guava:23.0',
+            'net.sf.jasperreports:jasperreports:6.5.1',
+            'commons-io:commons-io:2.6'
+
+    testCompile 'junit:junit:4.12'
+}
