+++ src/main/java/com/microting/report/jasper/ReportExportException.java
@@ -1,0 +1,10 @@
+package com.microting.report.jasper;
+
+public class ReportExportException extends Exception {
+
+	private static final long serialVersionUID = -1129692292804911374L;
+
+	public ReportExportException(Throwable e) {
+		super(e);
+	}
+}
