+++ src/main/java/com/microting/report/jasper/convertion/ReportConversionException.java
@@ -1,0 +1,10 @@
+package com.microting.report.jasper.convertion;
+
+public class ReportConversionException extends Exception {
+
+	private static final long serialVersionUID = 3612660591827286357L;
+
+	public ReportConversionException(Throwable e) {
+		super(e);
+	}
+}
